+++ apps/namadillo/src/App/Governance/LiveGovernanceProposals.tsx
@@ -5,7 +5,10 @@ import BigNumber from "bignumber.js";
 import clsx from "clsx";
 import { GoInfo } from "react-icons/go";
 import { generatePath, useNavigate } from "react-router-dom";
-import { secondsToDateTimeString } from "utils";
+import {
+  secondsToDateTimeString,
+  secondsToFullDateTimeString,
+} from "utils/dates";
 import { StatusLabel, TypeLabel, VotedLabel } from "./ProposalLabels";
 import { colors } from "./types";
 
@@ -47,7 +50,10 @@ const ProposalListItem: React.FC<{
     >
       <div className="flex items-center justify-between gap-4">
         <StatusLabel className="text-[10px] min-w-38" status={status} />
-        <div className="text-xs text-neutral-400">
+        <div
+          className="text-xs text-neutral-400"
+          title={secondsToFullDateTimeString(proposal.endTime)}
+        >
           Voting End on {secondsToDateTimeString(proposal.endTime)}&nbsp;&mdash;
           Epoch {proposal.endEpoch.toString()}
         </div>
